Fix migrations

+++ Modules/Airplane/Database/migrations/2021_12_26_114649_create_airplanes_table.php
@@ -17,7 +17,7 @@ class CreateAirplanesTable extends Migration
             $table->uuid('id')->primary();
             $table->string('name');
             $table->text('details');
-            $table->integer('code')->unique();
+            $table->unsignedBigInteger('code')->unique();
             $table->boolean('floor');
             $table->timestamps();
         });
