+++ Modules/Flight/Database/migrations/2021_12_26_113443_create_flights_table.php
@@ -15,7 +15,7 @@ class CreateFlightsTable extends Migration
     {
         Schema::create('flights', function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->integer('flight_number')->unique();
+            $table->unsignedBigInteger('flight_number')->unique();
             $table->time('takeoff');
             $table->time('landing');
             $table->string('source');
